Chore/board view (#33)
* Calendar and list refinement.

Made some refinement to the list and calendar modals. Accidentally found and fixed a bug when filtering tasks.

* Google handler

Added a google handler to reuse the login and logout logic. Renamed task folder to model.

* Added loader element to google actions

* Syncing with google

Synchronization when importing events from google without duplicates fixed!

* localizer start week day

* First approach to calendar preferences

* Week start day done

* bug on synchronizing events fixed

* toc

* first version to a board

* First prorotype for board view

* Board view

There are still some remaining fixes but board view prototype is finished.

* Board view + GitHub

- Board view is finished (only the refinement left). Removed the option of creating new tasks from the board so there's only one place in the app from which the creation of new tasks is allowed.
- GitHub is now connected and the synchronization of issues is available. A new router for github routes and a new context have been created.
When importing issues, a new list for each repo will be created, even though this can be considered to be an editable option from the preferences panel.

* Added icon to button